Remove old HDF interfaces (#845)
* Handle geometry properly in custom assembler tests.

* Update expression eval.

* Updates for bouning box tree

* Updates (hack) for geometry handling.

* Small fix.

* Remove some comments.

* Remove old-new XDMF mesh value collection code.

* Remove a lot of HDF5 code.

* Doc fix.

* Fix includes.

* Remove old tests.
