+++ cpp/dolfinx/io/HDF5File.h
@@ -7,12 +7,8 @@
 #pragma once
 
 #include "HDF5Interface.h"
+#include <Eigen/Dense>
 #include <dolfinx/common/MPI.h>
-#include <dolfinx/function/Function.h>
-#include <dolfinx/la/PETScVector.h>
-#include <dolfinx/mesh/Mesh.h>
-#include <dolfinx/mesh/MeshFunction.h>
-#include <dolfinx/mesh/MeshValueCollection.h>
 #include <memory>
 #include <string>
 #include <utility>
@@ -20,16 +16,6 @@
 
 namespace dolfinx
 {
-namespace la
-{
-class PETScVector;
-}
-
-namespace function
-{
-// class Function;
-class FunctionSpace;
-} // namespace function
 
 namespace io
 {
@@ -54,122 +40,6 @@ public:
   /// Flush buffered I/O to disk
   void flush();
 
-  /// Write points to file
-  void write(const std::vector<Eigen::Vector3d>& points,
-             const std::string name);
-
-  /// Write simple vector of double to file
-  void write(const std::vector<double>& values, const std::string name);
-
-  /// Write Vector to file in a format suitable for re-reading
-  void write(const la::PETScVector& x, const std::string name);
-
-  /// Read vector from file and optionally re-use any partitioning that
-  /// is available in the file
-  la::PETScVector read_vector(MPI_Comm comm, const std::string dataset_name,
-                              const bool use_partition_from_file) const;
-
-  /// Write Mesh to file in a format suitable for re-reading
-  void write(const mesh::Mesh& mesh, const std::string name);
-
-  /// Write Mesh of given cell dimension to file in a format
-  /// suitable for re-reading
-  void write(const mesh::Mesh& mesh, const int cell_dim,
-             const std::string name);
-
-  /// Write function::Function to file in a format suitable for re-reading
-  void write(const function::Function& u, const std::string name);
-
-  /// Write function::Function to file with a timestamp
-  void write(const function::Function& u, const std::string name,
-             double timestamp);
-
-  /// Read function::Function from file and distribute data according to the
-  /// mesh::Mesh and dofmap associated with the function::Function.  If the
-  /// 'name'
-  /// refers to a HDF5 group, then it is assumed that the function::Function
-  /// data is stored in the datasets within that group.  If the
-  /// 'name' refers to a HDF5 dataset within a group, then it is
-  /// assumed that it is a Vector, and the function::Function will be filled
-  /// from that Vector
-  function::Function read(std::shared_ptr<const function::FunctionSpace> V,
-                          const std::string name) const;
-
-  /// Read mesh::Mesh from file, using attribute data (e.g., cell type)
-  /// stored in the HDF5 file. Optionally re-use any partition data
-  /// in the file. This function requires all necessary data for
-  /// constructing a mesh::Mesh to be present in the HDF5 file.
-  mesh::Mesh read_mesh(const std::string data_path,
-                       bool use_partition_from_file,
-                       const mesh::GhostMode ghost_mode) const;
-
-  /// Read in the data from the first mesh in HDF file
-  ///
-  /// @param[in] data_path Path to topology and geometry datasets
-  /// @return Cell type, geometric points on each process, cell topology
-  ///   (in global vertex indexing), cell partition rank from file (if
-  ///   provided)
-  std::tuple<
-      mesh::CellType,
-      Eigen::Array<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>,
-      Eigen::Array<std::int64_t, Eigen::Dynamic, Eigen::Dynamic,
-                   Eigen::RowMajor>,
-      std::vector<std::int64_t>, std::vector<std::int64_t>>
-  read_mesh_data(const std::string data_path) const;
-
-  /// Write mesh::MeshFunction to file in a format suitable for re-reading
-  void write(const mesh::MeshFunction<std::size_t>& meshfunction,
-             const std::string name);
-
-  /// Write mesh::MeshFunction to file in a format suitable for re-reading
-  void write(const mesh::MeshFunction<int>& meshfunction,
-             const std::string name);
-
-  /// Write mesh::MeshFunction to file in a format suitable for re-reading
-  void write(const mesh::MeshFunction<double>& meshfunction,
-             const std::string name);
-
-  /// Read mesh::MeshFunction from file
-  mesh::MeshFunction<std::size_t>
-  read_mf_size_t(std::shared_ptr<const mesh::Mesh> mesh,
-                 const std::string name) const;
-
-  /// Read mesh::MeshFunction from file
-  mesh::MeshFunction<int> read_mf_int(std::shared_ptr<const mesh::Mesh> mesh,
-                                      const std::string name) const;
-
-  /// Read mesh::MeshFunction from file
-  mesh::MeshFunction<double>
-  read_mf_double(std::shared_ptr<const mesh::Mesh> mesh,
-                 const std::string name) const;
-
-  /// Write mesh::MeshValueCollection to file
-  void write(const mesh::MeshValueCollection<std::size_t>& mesh_values,
-             const std::string name);
-
-  /// Write mesh::MeshValueCollection to file
-  void write(const mesh::MeshValueCollection<double>& mesh_values,
-             const std::string name);
-
-  /// Write mesh::MeshValueCollection to file
-  void write(const mesh::MeshValueCollection<bool>& mesh_values,
-             const std::string name);
-
-  /// Read mesh::MeshValueCollection from file
-  mesh::MeshValueCollection<std::size_t>
-  read_mvc_size_t(std::shared_ptr<const mesh::Mesh> mesh,
-                  const std::string name) const;
-
-  /// Read mesh::MeshValueCollection from file
-  mesh::MeshValueCollection<double>
-  read_mvc_double(std::shared_ptr<const mesh::Mesh> mesh,
-                  const std::string name) const;
-
-  /// Read mesh::MeshValueCollection from file
-  mesh::MeshValueCollection<bool>
-  read_mvc_bool(std::shared_ptr<const mesh::Mesh> mesh,
-                const std::string name) const;
-
   /// Check if dataset exists in HDF5 file
   bool has_dataset(const std::string dataset_name) const;
 
@@ -185,49 +55,22 @@ public:
   /// Chunking parameter - partition data into fixed size blocks for efficiency
   bool chunking = false;
 
-private:
-  // Friend
-  friend class XDMFFile;
-
-  // Write a mesh::MeshFunction to file
-  template <typename T>
-  void write_mesh_function(const mesh::MeshFunction<T>& meshfunction,
-                           const std::string name);
-
-  // Read a mesh::MeshFunction from file
-  template <typename T>
-  mesh::MeshFunction<T>
-  read_mesh_function(std::shared_ptr<const mesh::Mesh> mesh,
-                     const std::string name) const;
-
-  // Write a mesh::MeshValueCollection to file (new version using vertex
-  // indices)
-  template <typename T>
-  void
-  write_mesh_value_collection(const mesh::MeshValueCollection<T>& mesh_values,
-                              const std::string name);
-
-  // Read a mesh::MeshValueCollection from file
-  template <typename T>
-  mesh::MeshValueCollection<T>
-  read_mesh_value_collection(std::shared_ptr<const mesh::Mesh> mesh,
-                             const std::string name) const;
-
-  // Write contiguous data to HDF5 data set. Data is flattened into
-  // a 1D array, e.g. [x0, y0, z0, x1, y1, z1] for a vector in 3D
+  /// Write contiguous data to HDF5 data set. Data is flattened into a
+  /// 1D array, e.g. [x0, y0, z0, x1, y1, z1] for a vector in 3D
   template <typename T>
   void write_data(const std::string dataset_name, const std::vector<T>& data,
                   const std::vector<std::int64_t>& global_size,
                   bool use_mpi_io);
 
-  // Write 2D dataset to HDF5. Eigen::Arrays on each process must have the same
-  // number of columns.
+  /// Write 2D dataset to HDF5. Eigen::Arrays on each process must have
+  /// the same number of columns.
   template <typename T>
   void write_data(
       const std::string dataset_name,
       Eigen::Array<T, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>& data,
       bool use_mpi_io);
 
+private:
   // HDF5 file descriptor/handle
   hid_t _hdf5_file_id;
